pyproject.toml
[project]
name = "final_login"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "oddsummer56", email = "oddsummer56@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.32.1",
    "pydantic>=2.10.3",
    "pymongo>=4.9.2",
    "motor>=3.6.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "pyjwt>=2.10.1",
    "starlette>=0.41.3",
    "itsdangerous>=2.2.0",
    "dnspython>=2.7.0",
    "httpx>=0.28.1",
    "kafka-python>=2.0.2",
    "boto3>=1.35.82",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pytest.ini_options]
pythonpath = [".", "src"]